Updates Ember to 1.10, Ember-CLI to 0.2.0 beta
Updates the addon structure to match convention

Lays groundwork to extend async-button to support the following format
```hbs
{{async-button arg1 arg2 action="save" default="Save" pending="Saving..."}}
{{! if possible, will try to do the following}}
{{async-button "save" arg1 arg2 default="Save" pending="Saving..."}}
```
