+++ .editorconfig
@@ -29,5 +29,5 @@ indent_size = 2
 indent_style = space
 indent_size = 2
 
-[*.md]
+[*.{diff,md}]
 trim_trailing_whitespace = false
