+++ .travis.yml
@@ -7,6 +7,10 @@ cache:
   directories:
     - node_modules
 
+before_install:
+  - "npm config set spin false"
+  - "npm install -g npm@^2"
+
 install:
   - npm install -g bower
   - npm install
