+++ index.js
@@ -1,3 +1,6 @@
+/* jshint node: true */
+'use strict';
+
 module.exports = {
   name: 'ember-cli-async-button'
 };
