+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "ember-cli-async-button",
   "version": "0.3.0",
-  "version": "0.2.0",
   "description": "When running async actions ensuring disabling of the button, re-enabling, and handling promise rejections is pretty boilerplate. This component packages up that behavior.",
   "directories": {
     "doc": "doc",
@@ -12,23 +11,28 @@
     "build": "ember build",
     "test": "ember test"
   },
-  "repository": "https://github.com/dockyard/ember-cli-async-button",
+  "repository": "",
   "engines": {
     "node": ">= 0.10.0"
   },
-  "author": "Brian Cardarella",
+  "author": "",
   "license": "MIT",
+  "dependencies": {
+    "ember-cli-htmlbars": "0.7.4"
+  },
   "devDependencies": {
-    "body-parser": "^1.2.0",
-    "broccoli-asset-rev": "0.3.0",
-    "broccoli-ember-hbs-template-compiler": "^1.6.1",
-    "ember-cli": "0.1.0",
-    "ember-cli-content-security-policy": "0.2.0",
+    "broccoli-asset-rev": "^2.0.0",
+    "ember-cli": "0.2.0-beta.1",
+    "ember-cli-babel": "^4.0.0",
+    "ember-cli-app-version": "0.3.1",
+    "ember-cli-content-security-policy": "0.3.0",
+    "ember-cli-dependency-checker": "0.0.7",
     "ember-cli-ic-ajax": "0.1.1",
-    "ember-cli-inject-live-reload": "^1.2.2",
-    "ember-cli-qunit": "0.1.0",
-    "ember-data": "1.0.0-beta.10",
-    "ember-cli-pretender": "^0.2.3",
+    "ember-cli-inject-live-reload": "^1.3.0",
+    "ember-cli-qunit": "0.3.8",
+    "ember-cli-uglify": "1.0.1",
+    "ember-data": "1.0.0-beta.15",
+    "ember-export-application-global": "^1.0.2",
     "express": "^4.8.5",
     "glob": "^4.0.5"
   },
