+++ tests/dummy/app/index.html
@@ -11,9 +11,15 @@
 
     <link rel="stylesheet" href="assets/vendor.css">
     <link rel="stylesheet" href="assets/dummy.css">
+
+    {{content-for 'head-footer'}}
   </head>
   <body>
+    {{content-for 'body'}}
+
     <script src="assets/vendor.js"></script>
     <script src="assets/dummy.js"></script>
+
+    {{content-for 'body-footer'}}
   </body>
 </html>
